+++ tests/test_resource.py
@@ -1,11 +1,16 @@
 import pytest
 from fastapi.routing import APIRouter
+from fastapi.security import SecurityScopes
 
+from clean_python import PermissionDenied
 from clean_python.fastapi import APIVersion
+from clean_python.fastapi import ctx
 from clean_python.fastapi import get
 from clean_python.fastapi import Resource
 from clean_python.fastapi import Stability
 from clean_python.fastapi import v
+from clean_python.fastapi.resource import check_scope_dependable
+from clean_python.oauth2 import Claims
 
 
 def test_subclass():
@@ -54,6 +59,7 @@ def test_get_router():
     assert route.name == "v1/get_test"
     assert route.tags == ["testing"]
     assert route.methods == {"GET"}
+    assert len(route.dependencies) == 0
     # 'self' is missing from the parameters
     assert list(route.param_convertors.keys()) == ["id"]
 
@@ -134,3 +140,43 @@ def test_get_less_stable_no_subclass():
 
     with pytest.raises(RuntimeError):
         resources[v(1)].get_less_stable(resources)
+
+
+def test_get_router_with_scope():
+    class TestResource(Resource, version=v(1), name="testing"):
+        @get("/foo/{id}", scope="foo")
+        def get_test(self, id: int):
+            return "ok"
+
+    resource = TestResource()
+
+    router = resource.get_router(v(1))
+
+    assert len(router.routes) == 1
+
+    route = router.routes[0]
+    (security,) = route.dependencies
+    assert security.scopes == ["foo"]
+    assert security.dependency is check_scope_dependable
+
+
+@pytest.mark.parametrize(
+    "ctx_scope, security_scopes, ok",
+    [
+        (["view"], ["view"], True),
+        ([], ["view"], False),
+        (["start"], ["view"], False),
+        (["view"], [], True),
+        (["view", "start"], ["view"], True),
+    ],
+)
+async def test_check_scope_dependable(ctx_scope, security_scopes, ok):
+    ctx.claims = Claims(
+        user={"id": "abc", "name": "jan"}, tenant=None, scope=frozenset(ctx_scope)
+    )
+
+    if ok:
+        await check_scope_dependable(SecurityScopes(security_scopes))
+    else:
+        with pytest.raises(PermissionDenied):
+            await check_scope_dependable(SecurityScopes(security_scopes))